pci: Reduce log level of error message
If a link doesn't train, we can end up with error messages like this:

[   63.027261959,3] PHB#0032[8:2]: LINK: Timeout waiting for electrical link
[   63.027265573,3] PHB#0032:00:00.0 Error -6 resetting

The first message is useful but the second message is just debug from
the core PCI code and is confusing to print to the console.

This reduces the second print to debug level so it's not seen by the
console by default.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Reported-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>