npu2: Remove DD1 support
Major changes in the NPU between DD1 and DD2 necessitated a fair bit of
revision-specific code.

Now that all our lab machines are DD2, we no longer test anything on DD1
and it's time to get rid of it.

Remove DD1-specific code and abort probe if we're running on a DD1 machine.

Cc: Alistair Popple <alistair@popple.id.au>
Cc: Reza Arbab <arbab@linux.vnet.ibm.com>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Acked-By: Alistair Popple <alistair@popple.id.au>
Acked-by: Reza Arbab <arbab@linux.vnet.ibm.com>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>