Disable stop states from OPAL
On ZZ, stop4,5,11 are enabled for PHYP, even though doing
so may cause problems with OPAL due to bugs in hcode.

For other platforms, this isn't so much of an issue as
we can just control stop states by the MRW. However the
rebuild-the-world approach to changing values there is a bit
annoying if you just want to rule out a specific stop state
from being problematic.

Provide an nvram option to override what's disabled in OPAL.

The OPAL mask is currently ~0xE0000000 (i.e. all but stop 0,1,2)

You can set an NVRAM override with:
  nvram -p ibm,skiboot --update-config opal-stop-state-disable-mask=0xFFFFFFF
This nvram override will disable *all* stop states.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>