mambo/mambo_utils.tcl: Inject an MCE at a specified address
Currently we don't support injecting an MCE on a specific address.
This is useful for testing functionality like memcpy_mcsafe()
(see https://patchwork.ozlabs.org/cover/893339/)

The core of the functionality is a routine called
inject_mce_ue_on_addr, which takes an addr argument and injects
an MCE (load/store with UE) when the specified address is accessed
by code. This functionality can easily be enhanced to cover
instruction UE's as well.

A sample use case to create an MCE on stack access would be

set addr [mysim display gpr 1]
inject_mce_ue_on_addr $addr

This would cause an mce on any r1 or r1 based access

Signed-off-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>