p9dsu: detect p9dsu variant even when hostboot doesn't tell us
The SuperMicro BMC can tell us what riser type we have, which dictates
the PCI slot tables. Usually, in an environment that a customer would
experience, Hostboot will do the query with an SMC specific patch
(not upstream as there's no platform specific code in hostboot)
and skiboot knows what variant it is based on the compatible string.

However, if you're using upstream hostboot, you only get the bare
'p9dsu' compatible type. We can work around this by asking the BMC
ourselves and setting the slot table appropriately. We do this
syncronously in platform init so that we don't start probing
PCI before we setup the slot table.

This adds a bit of funky logic in the p9dsu platform file, but on
the whole makes things simpler.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>