tests: Specfiy Qemu with a single environment variable
This allows Qemu to work out of the box if the user has an appropriate
Qemu installed.

Support for running TCG power guests has come a long way from when we
first added these test scripts. It makes sense to use the system Qemu
where possible, as for most people this will succeed.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>