hmi: Clear unknown debug trigger
On some systems, seeing hangs like this when Linux starts:
  [ 170.027252763,5] OCC: All Chip Rdy after 0 ms
  [ 170.062930145,5] INIT: Starting kernel at 0x20011000, fdt at 0x30ae0530 366247 bytes)
  [ 171.238270428,5] OPAL: Switch to little-endian OS

If you look at the in memory skiboot console (or do 'nvram -p
ibm,skiboot --update-config log-level-driver=7') we see the console get
spammed with:
  [ 5209.109790675,7] HMI: Received HMI interrupt: HMER = 0x0000400000000000
  [ 5209.109792716,7] HMI: Received HMI interrupt: HMER = 0x0000400000000000
  [ 5209.109794695,7] HMI: Received HMI interrupt: HMER = 0x0000400000000000
  [ 5209.109796689,7] HMI: Received HMI interrupt: HMER = 0x0000400000000000

We're taking the debug trigger (bit 17) early on, before the
hmi_debug_trigger function in the kernel is set up.

This clears the HMI in Skiboot and reports to the kernel instead of
bringing down the machine.

Signed-off-by: Ryan Grimm <grimm@linux.vnet.ibm.com>
Signed-off-by: Michael Neuling <mikey@neuling.org>
Reviewed-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>